requirements.txt
mlx-lm>=0.22.0
mlx>=0.22.0
transformers>=4.39.3
langchain-core>=0.1.0
langchain>=0.1.0
requests>=2.31.0
pydantic>=2.0.0
mlx-embedding-models>=0.0.5
lancedb>=0.3.0
pypdf>=3.15.1
python-docx>=0.8.11
psutil>=5.9.0
pandas>=2.0.0
numpy>=1.24.0
